Add ServiceLayer for Alamofire request
