+++ MercadoPagoSDK/MercadoPagoSDK/Services/Services/MercadoPagoService.swift
@@ -58,33 +58,18 @@ internal class MercadoPagoService: NSObject {
 
         UIApplication.shared.isNetworkActivityIndicatorVisible = true
 
-        #if PX_PRIVATE_POD
-            MercadoPagoSDKV4.request(request).responseData { response in
-                MercadoPagoService.debugPrint(response: response)
-                UIApplication.shared.isNetworkActivityIndicatorVisible = false
-                if let data = response.result.value, response.error == nil {
-                    success(data)
-                } else if let error = response.error as NSError? {
-                    failure?(error)
-                } else {
-                    let error: NSError = NSError(domain: "com.mercadopago.sdk", code: NSURLErrorCannotDecodeContentData, userInfo: nil)
-                    failure?(error)
-                }
+        PXServiceLayer().request(request).responseData { response in
+            MercadoPagoService.debugPrint(response: response)
+            UIApplication.shared.isNetworkActivityIndicatorVisible = false
+            if let data = response.result.value, response.error == nil {
+                success(data)
+            } else if let error = response.error as NSError? {
+                failure?(error)
+            } else {
+                let error: NSError = NSError(domain: "com.mercadopago.sdk", code: NSURLErrorCannotDecodeContentData, userInfo: nil)
+                failure?(error)
             }
-        #else
-            MercadoPagoSDKV4.request(request).responseData { response in
-                MercadoPagoService.debugPrint(response: response)
-                UIApplication.shared.isNetworkActivityIndicatorVisible = false
-                if let data = response.result.value, response.error == nil {
-                    success(data)
-                } else if let error = response.error as NSError? {
-                    failure?(error)
-                } else {
-                    let error: NSError = NSError(domain: "com.mercadopago.sdk", code: NSURLErrorCannotDecodeContentData, userInfo: nil)
-                    failure?(error)
-                }
-            }
-        #endif
+        }
     }
 }
